refactor : Move directory in service package
