+++ writer-service/src/main/java/com/dasd412/api/writerservice/adapter/in/security/filter/JwtAuthenticationFilter.java
@@ -6,7 +6,7 @@ import com.dasd412.api.writerservice.adapter.in.security.exception.LoginBadReque
 import com.dasd412.api.writerservice.adapter.in.security.dto.LoginRequestDTO;
 import com.dasd412.api.writerservice.adapter.in.security.jwt.JWTTokenProvider;
 import com.dasd412.api.writerservice.adapter.out.web.cookie.CookieProvider;
-import com.dasd412.api.writerservice.application.service.cache.refresh.RefreshTokenService;
+import com.dasd412.api.writerservice.application.service.security.refresh.RefreshTokenService;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.slf4j.Logger;
