+++ writer-service/src/main/java/com/dasd412/api/writerservice/application/service/writer/SaveWriterService.java
@@ -1,7 +1,6 @@
 package com.dasd412.api.writerservice.application.service.writer;
 
-import com.dasd412.api.writerservice.application.service.vo.AuthenticationVO;
-import com.dasd412.api.writerservice.domain.writer.Writer;
+import com.dasd412.api.writerservice.application.service.security.vo.AuthenticationVO;
 
 import java.util.concurrent.TimeoutException;
