+++ writer-service/src/main/java/com/dasd412/api/writerservice/common/config/security/SecurityConfiguration.java
@@ -3,7 +3,7 @@ package com.dasd412.api.writerservice.common.config.security;
 import com.dasd412.api.writerservice.adapter.in.security.filter.JwtAuthenticationFilter;
 import com.dasd412.api.writerservice.adapter.in.security.jwt.JWTTokenProvider;
 import com.dasd412.api.writerservice.adapter.out.web.cookie.CookieProvider;
-import com.dasd412.api.writerservice.application.service.cache.refresh.RefreshTokenService;
+import com.dasd412.api.writerservice.application.service.security.refresh.RefreshTokenService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
